Updated: #1 parsing process and geojson creation & refactoring
The osm parser can also handle relation objects now. Added mapnik module to gitignore, as it cant be uploaded due to long path problems. Removed unused node_modules. The map.png is now created by mapnik from the test.geojson, but it needs to be reduced to a certain tile boundary, as it is using simply all data right now.